+++ CHANGELOG.md
@@ -9,6 +9,7 @@ project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [0.1.0] - Unreleased
 ### Added
-- `toc` plugin
+- `toc` plugin.
+- `title` plugin.
 
 [0.1.0]: https://github.com/lumeland/markdown-plugins/releases/tag/v0.1.0